+++ CHANGELOG.md
@@ -474,6 +474,10 @@ $$Pow({x}, {n})=Pow\left({x}^2, \dfrac{n}{2}\right) * Pow\left(x, n \mod 2\right
 
 - Recursion
 
+## 0099 - Recover Binary Search Tree
+
+- Scan twice: from head, from tail
+
 ## 0415 - Add strings
 
 - Full adder